Legger validering av brevmottakere ut i egen fil, fikser opp i kode etter tilbakemelding
