+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/brev/mottaker/BrevmottakerService.kt
@@ -110,28 +110,4 @@ class BrevmottakerService(
             poststed = brevmottaker.poststed,
             landkode = brevmottaker.landkode,
         )
-
-    fun erBrevmottakerGyldig(brevmottaker: ManuellBrevmottaker): Boolean {
-        if (brevmottaker.landkode == "NO") {
-            return brevmottaker.navn.isNotEmpty() &&
-                brevmottaker.adresselinje1.isNotEmpty() &&
-                brevmottaker.postnummer.isNotEmpty() &&
-                brevmottaker.poststed.isNotEmpty()
-        } else {
-            // Utenlandske manuelle brevmottakere skal ha postnummer og poststed satt i adresselinjene
-            return brevmottaker.navn.isNotEmpty() &&
-                brevmottaker.adresselinje1.isNotEmpty() &&
-                brevmottaker.postnummer.isEmpty() &&
-                brevmottaker.poststed.isEmpty()
-        }
-    }
-
-    fun erBrevmottakereGyldige(brevmottakere: List<ManuellBrevmottaker>): Boolean {
-        brevmottakere.forEach {
-            if (!erBrevmottakerGyldig(it)) {
-                return false
-            }
-        }
-        return true
-    }
 }
